Tests to support implementation.
Created tests (and fixture classes) to grant the implementation continue
to work.
There are some tests I was not able to make, by now.

Removes some declarations on `composer.json` that were not needed and
added a better description and keywords to ease the finding of the
project on Packagist.
Added also PHPUnit on development requirements to ease the tests without
Respect/Foundation.